JBIDE-16135 EL Content Assist in Java Editor: Make Content Assistant Additional Info window to support URLs.
Content Assist Additional Infos and Tooltips (Hovers) were made to be like in Java Content Assist (Support links, Back/Forward/Open actions, icons and so on)
The new	look works equally in JST editors and in Java Editor (for EL in	Java Strings).
The following JUnit Tests were updated (in another PR) for the changes made to CA Add. Info's and Tooltip's (Hover's) content:
org.jboss.tools.jsf.jsp.ca.test.CAForELJavaAndJSTCompareTest
org.jboss.tools.jsf.jsp.ca.test.CAJsfAddInfoInELMessagesTest
org.jboss.tools.jsf.jsp.hover.ELTooltipTest
