--- plugins/org.jboss.tools.jst.web.ui/src/org/jboss/tools/jst/web/ui/internal/editor/contentassist/messages.properties
@@ -1,9 +1,0 @@
-NO_JAVADOC=No Javadoc could be found<br>
-JspEditorPlugin_additionalInfo_affordance=Press 'Tab' from proposal table or click for focus
-JavadocHover_back=Back
-JavadocHover_back_toElement_toolTip=Back to {0}
-JavadocHover_forward= Forward
-JavadocHover_forward_toElement_toolTip=Forward to {0}
-JavadocHover_forward_toolTip=Forward
-JavadocHover_openDeclaration= Open Declaration
-JavadocHover_showInJavadoc= Show in Javadoc View
